pyproject.toml
[project]
name = "pycharmmiscproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.6",
    "python-multipart>=0.0.20",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
]
